% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_preference.R
\name{abs_preference}
\alias{abs_preference}
\title{# Break data into discrete categories for coloring-by in histogram. Returns a vector with one of three
labels: "Pro-Black Preference"; "No Preference"; "Pro-White Preference"}
\usage{
abs_preference(iatscore)
}
\arguments{
\item{iatscore}{IAT score}
}
\description{
# Break data into discrete categories for coloring-by in histogram. Returns a vector with one of three
labels: "Pro-Black Preference"; "No Preference"; "Pro-White Preference"
}
\examples{
abs_preference()
}
\keyword{iat}
\keyword{implicit}
