% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abspref3.R
\name{abspref3}
\alias{abspref3}
\title{# Break data into discrete categories for coloring-by in histogram. Returns a vector with one of three
labels: "Pro-Black Preference"; "No Preference"; "Pro-White Preference"}
\usage{
abspref3(iatscore)
}
\arguments{
\item{iatscore}{IAT score}
}
\description{
# Break data into discrete categories for coloring-by in histogram. Returns a vector with one of three
labels: "Pro-Black Preference"; "No Preference"; "Pro-White Preference"
}
\examples{
abspref3()
}
\keyword{iat}
\keyword{implicit}
