% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abspref6.R
\name{abspref6}
\alias{abspref6}
\title{# Break data into discrete categories for coloring-by in histogram. Returns a vector with one of six labels}
\usage{
abspref6(iatscore)
}
\arguments{
\item{iatscore}{IAT score}
}
\description{
# Break data into discrete categories for coloring-by in histogram. Returns a vector with one of six labels
}
\examples{
abspref6()
}
\keyword{iat}
\keyword{implicit}
