% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correffectsize.R
\name{correffectsize}
\alias{correffectsize}
\title{Define function that prints size of correlation coefficients}
\usage{
correffectsize(corrtestvalue)
}
\arguments{
\item{corrtestvalue}{Exact correlation value}
}
\description{
Define function that prints size of correlation coefficients
}
\examples{
correffectsize()

}
\keyword{correlation}
\keyword{iat}
\keyword{implicit}
