% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrsig.R
\name{corrsig}
\alias{corrsig}
\title{A signifcance function}
\usage{
corrsig(corrsigvalue)
}
\arguments{
\item{corrsigvalue}{Exact significance level}
}
\description{
Conditionally prints correlation significance level
}
\examples{
corrsig()
}
\keyword{correlation}
\keyword{iat}
\keyword{implicit}
