% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participants.R
\name{participants}
\alias{participants}
\title{# participants}
\usage{
participants(data)
}
\arguments{
\item{NULL}{no parameters}
}
\description{
# participants
}
\examples{
participants()
}
\keyword{participants}
