% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participants_asglue.R
\name{participants_asglue}
\alias{participants_asglue}
\title{# participants_asglue}
\usage{
participants_asglue(data)
}
\arguments{
\item{NULL}{no parameters}
}
\description{
# participants_asglue
}
\examples{
participants_asglue()
}
\keyword{participants}
